+++ python/GafferImageUI/CatalogueUI.py
@@ -579,7 +579,7 @@ def _duplicateImages( catalogue, plugIndices ) :
 		orderedImages.insert( uiInsertionIndex, imageCopy )
 		insertions.append( len(images) - 1 )
 
-	_ImagesPath._updateUIIndices( orderedImages )
+	_ImagesPath._reorderImages( orderedImages )
 
 	return insertions
 
@@ -627,9 +627,12 @@ class _ImagesPath( Gaffer.Path ) :
 
 	def _orderedImages( self ) :
 
-		# Avoid repeat lookups for plugs with no ui index by first getting all
-		# images with their plug indices, then updating those with any metadata
+		# Start with the order of the Image plugs. Ideally this is all we want
+		# to use.
 		imageAndIndices = [ [ image, plugIndex ] for plugIndex, image in enumerate( self.__images.children() ) ]
+		# Apply ordering from legacy metadata. We used this before we had the
+		# ability to reorder the Image plugs themselves. We'll remove the metadata
+		# the first chance we get - when `_reorderImages()` is called.
 		for imageAndIndex in imageAndIndices :
 			uiIndex = Gaffer.Metadata.value( imageAndIndex[0], _ImagesPath.indexMetadataName )
 			if uiIndex is not None :
@@ -663,10 +666,16 @@ class _ImagesPath( Gaffer.Path ) :
 		}
 
 	@staticmethod
-	def _updateUIIndices( orderedImages ) :
+	def _reorderImages( orderedImages ) :
 
-		for i, image in enumerate( orderedImages ) :
-			Gaffer.Metadata.registerValue( image, _ImagesPath.indexMetadataName, i )
+		# Remove legacy metadata that was once used to reorder images before we
+		# had a `reorderChildren()` method.
+		for image in orderedImages :
+			Gaffer.Metadata.deregisterValue( image, _ImagesPath.indexMetadataName )
+
+		# Because we can reorder the images properly via the API now.
+		if orderedImages :
+			orderedImages[0].parent().reorderChildren( orderedImages )
 
 	def __childAdded( self, parent, child ) :
 
@@ -956,7 +965,7 @@ class ImageListing( GafferUI.PlugValueWidget ) :
 				self.__images().removeChild( image )
 				orderedImages.remove( image )
 
-			_ImagesPath._updateUIIndices( orderedImages )
+			_ImagesPath._reorderImages( orderedImages )
 
 			# Figure out new selection
 			if orderedImages :
@@ -1099,7 +1108,7 @@ class ImageListing( GafferUI.PlugValueWidget ) :
 			images.insert( newIndex, image )
 			previous = image
 
-		_ImagesPath._updateUIIndices( [image for image in images if image ] )
+		_ImagesPath._reorderImages( [image for image in images if image ] )
 
 		self.__pathListing.getPath().pathChangedSignal()( self.__pathListing.getPath() )